Moved the chart to a separate component
