+++ app/components/TimeNode/TimeNodeStatistics.js
@@ -1,10 +1,9 @@
 import React, { Component } from 'react';
 import { inject, observer } from 'mobx-react';
 import PropTypes from 'prop-types';
-import { Chart } from 'chart.js';
-
 import Alert from '../Common/Alert';
 import { TIMENODE_STATUS } from '../../stores/TimeNodeStore';
+import ExecutedGraph from './ExecutedGraph';
 
 @inject('timeNodeStore')
 @observer
@@ -14,8 +13,6 @@ class TimeNodeStatistics extends Component {
     this.state = {
       timeNodeDisabled: null
     };
-
-    this.refreshChart = this.refreshChart.bind(this);
     this.startTimeNode = this.startTimeNode.bind(this);
     this.stopTimeNode = this.stopTimeNode.bind(this);
   }
@@ -25,17 +22,6 @@ class TimeNodeStatistics extends Component {
     this.setState({
       timeNodeDisabled: this.props.timeNodeStore.nodeStatus === TIMENODE_STATUS.DISABLED
     });
-
-    this.refreshChart();
-  }
-
-  componentDidMount() {
-    // Refreshes the graph every 5 seconds
-    this.interval = setInterval(this.refreshChart, 5000);
-  }
-
-  componentWillUnmount() {
-    clearInterval(this.interval);
   }
 
   getStopButton() {
@@ -61,116 +47,6 @@ class TimeNodeStatistics extends Component {
 
   refreshStats() {
     this.props.timeNodeStore.updateStats();
-    this.refreshChart();
-  }
-
-  refreshChart() {
-    const data = this.props.timeNodeStore.executedTransactions;
-
-    let timeIntervals = [];
-
-    // This section sorts the executed transactions by hour into an array
-    for (let transaction of data) {
-      const transactionDate = new Date(transaction.timestamp);
-      transactionDate.setHours(transactionDate.getUTCHours(),0,0,0);
-
-      let timeIntervalExists = false;
-
-      // Check if that time interval already exists in the array
-      for (let i = 0; i < timeIntervals.length; i++) {
-        // If so, add the transaction to that array
-        if (timeIntervals[i].datetime.getTime() === transactionDate.getTime()) {
-          timeIntervals[i].executed += 1;
-          timeIntervalExists = true;
-        }
-      }
-
-      // If the time interval doesn't exist already
-      if (!timeIntervalExists) {
-        // Create the time interval and bump the executed flag
-        timeIntervals.push({ 'datetime': transactionDate, 'executed': 1 });
-      }
-    }
-
-    const ctx = this.chartRef.getContext('2d');
-    timeIntervals.sort(this.compareDates);
-
-    if (timeIntervals.length > 0) {
-      const dataLabels = timeIntervals.map(interval => interval.datetime.getHours() + ':00');
-      const dataExecuted = timeIntervals.map(interval => interval.executed);
-
-      new Chart(ctx, {
-        type: 'line',
-        data: {
-          labels: dataLabels,
-          datasets: [{
-            data: dataExecuted,
-            backgroundColor:'rgba(33, 255, 255, 0.2)',
-            borderColor: 'rgba(33, 255, 255, 1)',
-            borderWidth: 2,
-            datalabels: {
-              align: 'start',
-              anchor: 'start'
-            }
-          }]
-        },
-        options: {
-          animation: {
-            duration: 1,
-            onComplete: function () {
-              const chartInstance = this.chart;
-              const ctx = this.chart.ctx;
-              ctx.fillStyle = 'black';
-              ctx.textAlign = 'left';
-              ctx.textBaseline = 'bottom';
-
-              this.data.datasets.forEach(function (dataset, i) {
-                  const meta = chartInstance.controller.getDatasetMeta(i);
-                  meta.data.forEach(function (bar, index) {
-                      var data = dataset.data[index];
-                      ctx.fillText(data, bar._model.x, bar._model.y - 5);
-                  });
-              });
-            }
-          },
-          elements: {
-            line: {
-              tension: 0
-            }
-          },
-          legend: {
-            display: false
-          },
-          scales: {
-            yAxes: [{
-              display: false,
-              gridLines: {
-                tickMarkLength: 0
-              }
-            }],
-            xAxes: [{
-              display: true
-            }],
-          },
-          layout: {
-            padding: {
-              top: 20
-            }
-          }
-        }
-      });
-    } else {
-      ctx.clearRect(0, 0, this.chartRef.width, this.chartRef.height);
-      ctx.font = '20px Helvetica';
-      ctx.textAlign = 'center';
-      ctx.fillText('No data yet.', 150, 65);
-    }
-  }
-
-  compareDates(a, b) {
-    if (a.datetime.getTime() < b.datetime.getTime()) return -1;
-    if (a.datetime.getTime() > b.datetime.getTime()) return 1;
-    return 0;
   }
 
   render() {
@@ -227,8 +103,8 @@ class TimeNodeStatistics extends Component {
                   </ul>
                 </div>
               </div>
-              <div className="card-body no-padding">
-                <canvas id="myChart" ref={(el) => this.chartRef = el}></canvas>
+              <div ref={(el) => this.chartContainer = el} className="card-body no-padding">
+                {this.props.timeNodeStore.executedTransactions.length > 0 ? <ExecutedGraph /> : <p className="my-5 text-center">'No data yet.'</p>}
               </div>
             </div>
           </div>
